+++ src/theme/MDXComponents/Code/index.tsx
@@ -1,11 +1,37 @@
-import type { ComponentProps } from 'react';
-import React from 'react';
-import CopyButton from '@theme/CodeBlock/CopyButton';
-import type { Props } from '@theme/MDXComponents/Code';
-import codeBlockStyles from './CodeBlock.module.css';
-import clsx from 'clsx';
+/**
+ * Modified from https://github.com/facebook/docusaurus/blob/main/packages/docusaurus-theme-classic/src/theme/MDXComponents/Code.tsx
+ * 
+ * The MIT License (MIT)
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ * Copyright (c) 2024 MoonBit Open Course Team
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
 
-function shouldBeInline(props: Props): boolean {
+import type {ComponentProps} from 'react';
+import React from 'react';
+import CodeBlock from '../../CodeBlock';
+import CodeInline from '@theme/CodeInline';
+import type {Props} from '@theme/MDXComponents/Code';
+
+function shouldBeInline(props: Props) {
   return (
     // empty code blocks have no props.children,
     // see https://github.com/facebook/docusaurus/pull/9704
@@ -16,37 +42,10 @@ function shouldBeInline(props: Props): boolean {
   );
 }
 
-function getTextForCopy(node: React.ReactNode): string {
-  if (node === null) return '';
-
-  switch (typeof node) {
-    case 'string':
-    case 'number':
-      return node.toString();
-    case 'boolean':
-      return '';
-    case 'object':
-      if (node instanceof Array) return node.map(getTextForCopy).join('');
-      if ('props' in node) return getTextForCopy(node.props.children);
-    default:
-      return '';
-  }
-}
-
-function CodeBlock(props: ComponentProps<'code'>): JSX.Element {
-  const codeRef = React.useRef<HTMLElement>(null);
-  const code = getTextForCopy(props.children);
-
-  return (
-    <div className={codeBlockStyles.CodeBlock}>
-      <pre className={clsx(codeBlockStyles.pre, 'shiki')}>
-        <code {...props} ref={codeRef} />
-      </pre>
-      <CopyButton className={codeBlockStyles.button} code={code} />
-    </div>
+export default function MDXCode(props: Props): JSX.Element {
+  return shouldBeInline(props) ? (
+    <CodeInline {...props} />
+  ) : (
+    <CodeBlock {...(props as ComponentProps<typeof CodeBlock>)} />
   );
-}
-
-export default function MDXCode(props): JSX.Element {
-  return shouldBeInline(props) ? <code {...props} /> : <CodeBlock {...props} />;
 }
